Separate api function form file app.py by blueprint
